Fixed issue with ClassVar syft_internal_type
- Moved ActionDataEmpty out and use it to propagate syft_internal_type
