+++ packages/syft/src/syft/core/node/new/action_data_empty.py
@@ -1,0 +1,20 @@
+# future
+from __future__ import annotations
+
+# stdlib
+from typing import Any
+from typing import Optional
+from typing import Type
+
+# relative
+from ....core.node.common.node_table.syft_object import SYFT_OBJECT_VERSION_1
+from ....core.node.common.node_table.syft_object import SyftObject
+from ...common.serde.serializable import serializable
+
+
+@serializable(recursive_serde=True)
+class ActionDataEmpty(SyftObject):
+    __canonical_name__ = "ActionDataEmpty"
+    __version__ = SYFT_OBJECT_VERSION_1
+
+    syft_internal_type: Optional[Type] = Any
